Add PromptChange for holding multi-project changes
