+++ core/utils.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"slices"
 	"sort"
 	"strings"
 
@@ -308,7 +307,7 @@ func GetChanges(
 			return changes, err
 		}
 
-		if len(projectKey) > 0 && !slices.Contains(c.Projects, projectKey) {
+		if len(projectKey) > 0 && c.Project != projectKey {
 			continue
 		}
